requirements.txt
dj-database-url==0.4.0
Django==1.9.4
gunicorn==19.4.5
jsonfield==1.0.3
nltk==3.2
psycopg2==2.6.1
textblob==0.11.1
python-twitter
whitenoise==3.0
numpy==1.9.2
scipy==0.15.1
humanize==0.5.1
